refactor: move navbar link data to index ts
