+++ components/elements/Layout/Navbar/index.tsx
@@ -1,6 +1,47 @@
 import Image from "next/image";
 import Link from "next/link";
 import { NavigationButtons } from "./NavigationButtons";
+import { NavDropdown, NavLink } from "./interface";
+import { Home } from "@/components/icons/Home";
+import { Laptop } from "@/components/icons";
+import { DoorOpen, Pen, User2 } from "lucide-react";
+
+const NavbarLinks: (NavLink | NavDropdown)[] = [
+  {
+    href: "#",
+    label: "Home",
+    icon: (
+      <Home size="w-6 h-6" fill="fill-text-dark-1" className="stroke-[0.5]" />
+    ),
+  },
+  {
+    href: "#",
+    label: "Dashboard",
+    icon: (
+      <Laptop size="w-6 h-6" fill="fill-text-dark-1" className="stroke-[0.5]" />
+    ),
+  },
+  {
+    href: "#",
+    label: "Username",
+    icon: <User2 />,
+    children: [
+      {
+        href: "#",
+        label: "Edit Profile",
+        icon: <Pen />,
+        className: "hover:bg-[#FF9900] hover:text-text-light-3",
+      },
+      {
+        href: "#",
+        label: "Keluar",
+        icon: <DoorOpen />,
+        className:
+          "hover:bg-accents-red-1 hover:text-text-light-3 font-bold text-accents-red-1",
+      },
+    ],
+  },
+];
 
 export const Navbar = () => {
   return (
@@ -24,7 +65,10 @@ export const Navbar = () => {
             />
           </Link>
           <div className="flex items-center gap-3 md:gap-9">
-            <NavigationButtons className="hidden lg:flex" />
+            <NavigationButtons
+              className="hidden lg:flex"
+              navbarLinks={NavbarLinks}
+            />
           </div>
         </div>
         <Image
